create multiple files
